# docs/swagger.json
{
  "basePath": "/",
  "definitions": {
    "api.TransactionRequest": {
      "description": "Transaction request structure for invoking or evaluating chaincode",
      "properties": {
        "args": {
          "description": "Arguments to pass to the chaincode function",
          "example": [
            "[]"
          ],
          "items": {
            "type": "string"
          },
          "type": "array"
        },
        "chaincode_name": {
          "description": "Name of the chaincode to invoke",
          "example": "mycc",
          "type": "string"
        },
        "function": {
          "description": "Function name to call in the chaincode",
          "example": "GetAllAssets",
          "type": "string"
        }
      },
      "type": "object"
    },
    "api.TransactionResponse": {
      "description": "Response structure for chaincode transactions",
      "properties": {
        "block_number": {
          "description": "Block number where the transaction was committed",
          "example": 123,
          "type": "integer"
        },
        "error": {
          "description": "Error message (if failed)",
          "example": "Invalid arguments",
          "type": "string"
        },
        "result": {
          "description": "Result of the transaction (if successful)",
          "example": "{\"key\":\"value\"}",
          "type": "string"
        },
        "result_code": {
          "description": "Result code from the chaincode",
          "example": 200,
          "type": "integer"
        },
        "status": {
          "description": "Status of the transaction (\"success\" or \"error\")",
          "example": "success",
          "type": "string"
        },
        "success": {
          "description": "Whether the transaction was successful",
          "example": true,
          "type": "boolean"
        },
        "tx_id": {
          "description": "Transaction ID",
          "example": "tx123",
          "type": "string"
        }
      },
      "type": "object"
    }
  },
  "info": {
    "contact": {},
    "description": "API for interacting with Hyperledger Fabric network.\u003cbr/\u003e\u003cbr/\u003eAvailable chaincodes:\u003cbr/\u003e\u003ca href=\"/swagger/demochainlaunch.json\"\u003edemochainlaunch\u003c/a\u003e ",
    "title": "Hyperledger Fabric API",
    "version": "1.0"
  },
  "paths": {
    "/api/evaluate": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Evaluates a transaction on the Hyperledger Fabric network without committing it",
        "parameters": [
          {
            "description": "Transaction Request",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/api.TransactionRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/api.TransactionResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/api.TransactionResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/api.TransactionResponse"
            }
          }
        },
        "summary": "Evaluate a chaincode transaction",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/invoke": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Invokes a transaction on the Hyperledger Fabric network",
        "parameters": [
          {
            "description": "Transaction Request",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/api.TransactionRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/api.TransactionResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/api.TransactionResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/api.TransactionResponse"
            }
          }
        },
        "summary": "Invoke a chaincode transaction",
        "tags": [
          "transactions"
        ]
      }
    },
    "/refresh": {
      "get": {
        "description": "Reloads chaincode metadata and updates dynamic endpoints for all chaincodes",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/api.TransactionResponse"
            }
          }
        },
        "summary": "Refresh chaincode metadata and dynamic endpoints",
        "tags": [
          "admin"
        ]
      }
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0"
}